% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{conf_sample}
\alias{conf_sample}
\title{Generate a sample configuration file}
\usage{
conf_sample(file = stdout())
}
\arguments{
\item{file}{file path or a connection to write the configurations to.}
}
\value{
None
}
\description{
Generate a sample configuration file
}
