% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_url}
\alias{default_url}
\alias{default_host}
\alias{default_port}
\alias{default_protocol}
\title{Get or set default host, port, and protocols}
\usage{
default_host(host, allow0 = TRUE)

default_port(port)

default_protocol(protocol = c("", "http", "https"))
}
\arguments{
\item{host}{'IPv4' address}

\item{allow0}{internally used}

\item{port}{an integer of port to expose, ranging from 1 to 65535}

\item{protocol}{either 'http' or 'https'}
}
\value{
The current default options
}
\description{
Get or set default host, port, and protocols
}
\details{
By default, servers will run on local host ('127.0.0.1') at port
'7033' with protocol 'http'. The tasks will be sent to this address too.
}
