% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_tasks}
\alias{list_tasks}
\title{Query and list all your submitted tasks}
\usage{
list_tasks(
  status = c("valid", "running", "init", "finish", "all"),
  order = FALSE,
  expire = 0
)
}
\arguments{
\item{status}{filter task status on the server, choices are \code{'valid'},
\code{'init'} (submitted, waiting to run), \code{'running'} (running
task), \code{'finish'} (finished task), and \code{'canceled'} (canceled
by the server)}

\item{order}{whether to order by date submitted (in descending order);
default is false}

\item{expire}{positive number (in seconds) to filter out tasks
that have been submitted most recently, or 0 (default) to list tasks
regardless of their dates. For example, \code{expire=100} will only list
tasks submitted during the past 100 seconds.}
}
\value{
A data frame listing tasks submitted, columns are
\describe{
\item{\code{name}}{task name (ID)}
\item{\code{userid}}{your user ID}
\item{\code{submitted}}{1 if submitted and 0 otherwise}
\item{\code{collected}}{1 if result has been collected and 0 otherwise}
\item{\code{error}}{1 if error occurs and 0 otherwise}
\item{\code{path}}{the local directory that stores the task data}
\item{\code{serverip}}{server address if the task has been submitted}
\item{\code{serverport}}{server port if the task has been submitted}
\item{\code{removed}}{1 if the task has been removed and 0 otherwise}
\item{\code{time_added}}{UNIX time of time when task is created. Use
\code{as.POSIXct(time_added, origin="1970-01-01")} to convert to read-able
time; see \code{\link{as.POSIXct}}}
}
}
\description{
Query and list all your submitted tasks
}
