% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterFunctionsSocket2.R
\name{makeClusterFunctionsSocket2}
\alias{makeClusterFunctionsSocket2}
\title{Make 'batchtools' cluster with 'parallel' package}
\usage{
makeClusterFunctionsSocket2(ncpus = NA_integer_, fs.latency = 65)
}
\arguments{
\item{ncpus, fs.latency}{see \code{\link[batchtools]{makeClusterFunctionsSocket}}}
}
\value{
A \code{\link[batchtools]{makeClusterFunctions}}.
}
\description{
An implementation of \code{\link[batchtools]{makeClusterFunctionsSocket}}
with native 'parallel' package if 'snow' is not installed
}
