% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{queue_task}
\alias{queue_task}
\title{Add a task to a queue}
\usage{
queue_task(task, userid)
}
\arguments{
\item{task}{a task object}

\item{userid}{user ID}
}
\value{
None
}
\description{
called by server internally, not supposed to be called directly.
}
