% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{restbatch-db}
\alias{restbatch-db}
\alias{database_fix}
\alias{database_restore}
\title{Fix or restore database}
\usage{
database_fix()

database_restore(
  bkup,
  backup_current = TRUE,
  tables = c("user", "client-task", "server-task"),
  owner_key = c("drop_old", "keep_both")
)
}
\arguments{
\item{bkup}{path to the backup folder, should contains \code{'DB/restbatch.sqlite'}}

\item{backup_current}{whether to backup current one before restoring}

\item{tables}{which table(s) to restore}

\item{owner_key}{whether to keep the old owner key; default is \code{'drop_old'}}
}
\value{
None
}
\description{
These functions must run interactively.
}
