% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{restbatch-requests}
\alias{restbatch-requests}
\alias{request_server}
\alias{request_task_list}
\title{Send requests to the server to query the server status or task status}
\usage{
request_server(
  path,
  host = default_host(),
  port = default_port(),
  protocol = default_protocol(),
  body = list(),
  header = list(),
  method = c("POST", "GET"),
  encode = "json",
  timeout = 15
)

request_task_list(
  task_status = "valid",
  host = default_host(allow0 = FALSE),
  port = default_port(),
  protocol = default_protocol(),
  path = "task/list"
)
}
\arguments{
\item{protocol, host, port, path}{server location-related configurations, the
'url' of request will be \code{protocol://host:port/path}}

\item{body}{a list of request body}

\item{header}{additional header key-value pairs}

\item{method}{method of request; choices are \code{'POST'} and \code{'GET'}}

\item{encode}{serialization method to encode request body}

\item{timeout}{maximum waiting time in seconds before aborting the request;
default is 15. If the server is off or fails to respond before time running
out, the request will result in an error.}

\item{task_status}{task status to filter}
}
\value{
\code{'httr'} response. You can use \code{\link[httr]{content}} to
check the response contents.
}
\description{
\describe{
\item{\code{request_server}}{sends general request with authentication
tokens ready in the request headers}
\item{\code{request_task_list}}{get list of tasks of your current user ID
(\code{userid}) from the server}
}
}
