% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{restbench-handlers}
\alias{restbench-handlers}
\alias{run_task}
\alias{handler_unpack_task}
\alias{handler_query_task}
\alias{handler_validate_server}
\alias{handler_validate_auth}
\title{Default handlers to process incoming requests}
\usage{
run_task(task, userid)

handler_unpack_task(req)

handler_query_task(userid, status = "valid")

handler_validate_server(req)

handler_validate_auth(req, res)
}
\arguments{
\item{task}{a task instance}

\item{userid}{which user submits the task}

\item{req, res}{web request/response variables (see \code{'plumber'} package)}

\item{status}{task status to filter/query; choices are \code{'valid'},
\code{'init'} (submitted, waiting to run), \code{'running'} (running
task), \code{'finish'} (finished task), and \code{'canceled'} (canceled
by the server)}
}
\value{
\code{run_task} returns a named list with:
\describe{
\item{\code{task}}{task object created by internal \code{new_task} function}
\item{\code{userid}}{a 32-character string of user's ID}
\item{\code{future}}{a future object that schedules jobs asynchronously}.
}
\code{handler_unpack_task}, \code{handler_query_task} return task instances
created by the internal \code{new_task} function. Others return lists or
\code{res} that will be processed by the \code{'plumber'} package.
}
\description{
\describe{
\item{\code{run_task}}{schedules batch jobs and put them to a running list}
\item{\code{handler_unpack_task}}{unpack a request and convert to a task}
\item{\code{handler_query_task}}{obtain task status on the server}
\item{\code{handler_validate_server}}{handles validation request}
\item{\code{handler_validate_auth}}{a layer that handles authentication}
}
}
