% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{generate_service}
\alias{generate_service}
\title{Generate script to install the 'restbatch' as Linux service}
\usage{
generate_service(
  host = "127.0.0.1",
  port = 7033,
  save_settings_path = "~/.restbatch"
)
}
\arguments{
\item{save_settings_path}{where to save the settings file}
}
\description{
Generate script to install the 'restbatch' as Linux service
}
