% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{restbatch-requests}
\alias{restbatch-requests}
\alias{request_server}
\alias{request_task_list}
\title{Send requests to the server to query the server status or task status}
\usage{
request_server(
  url,
  body = list(),
  header = list(),
  method = c("POST", "GET"),
  encode = "json"
)

request_task_list(
  task_status = "valid",
  host = default_host(allow0 = FALSE),
  port = default_port(),
  protocol = default_protocol(),
  path = "jobs/list"
)
}
\arguments{
\item{url, protocol, host, port}{server location-related configurations}

\item{body}{a list of request body}

\item{header}{additional header key-value pairs}

\item{method}{method of request; choices are \code{'POST'} and \code{'GET'}}

\item{encode}{serialization method to encode request body}

\item{task_status}{task status to filter}

\item{path}{route path to send requests to}
}
\value{
\code{'httr'} response. You can use \code{\link[httr]{content}} to
check the response contents.
}
\description{
\describe{
\item{\code{request_server}}{sends general request with authentication
tokens ready in the request headers}
\item{\code{request_task_list}}{get list of tasks of your current userid
from the server}
}
}
